sinc function added

+++ ivy/functional/backends/tensorflow/elementwise.py
@@ -578,7 +578,7 @@ def sinc(
     *,
     out: Optional[Union[tf.Tensor, tf.Variable]] = None,
 ) -> Union[tf.Tensor, tf.Variable]:
-    return tf.experimental.numpy.sinc(x)
+    return tf.cast(tf.experimental.numpy.sinc(x), x.dtype)
 
 
 def sinh(
